migration updated to set double notes

+++ database/migrations/2021_02_03_113214_create_images_table.php
@@ -19,7 +19,7 @@ class CreateImagesTable extends Migration
             $table->integer('idUser')->nullable();
             $table->integer('idBroadcastList')->nullable();
             $table->string('legend')->nullable();
-            $table->float('note')->nullable();
+            $table->double('note', 8, 2)->nullable();
             $table->boolean('exifs')->nullable();
             $table->timestamps();
         });
